requirements.txt
python-dotenv

streamlit
streamlit-chat
streamlit-authenticator
streamlit_mic_recorder

gtts
openai
assemblyai
mistralai